requirements.txt
dotenv
requests
RPi.GPIO
paho
os
